+++ casa_formats_io/glue_factory.py
@@ -13,12 +13,15 @@ from casa_formats_io.casa_low_level_io.table import CASATable
 __all__ = ['read_spectral_cube', 'parse_spectral_cube']
 
 
-def is_casa_measurement_set(filename):
-    # TODO: we can do better than this, just a quick hack for now
-    return os.path.isdir(filename) and filename.endswith(('.ms', '.ms/'))
+def is_casa_table(filename):
+    if (isinstance(filename, str) and
+            os.path.isdir(filename) and
+            os.path.exists(os.path.join(filename, 'table.dat'))):
+        with open(os.path.join(filename, 'table.dat'), 'rb') as f:
+            return f.read(4) == b'\xbe\xbe\xbe\xbe'
 
 
-def table_to_glue_data(table, label_prefix, data_desc_id):
+def table_to_glue_data(table, label):
 
     # Glue can actually understand astropy tables, but doesn't work well with
     # some columns being vector columns, nor complex numbers, so we expand
@@ -26,6 +29,10 @@ def table_to_glue_data(table, label_prefix, data_desc_id):
 
     # Split out vector columns into single columns
     for colname in table.colnames:
+        # Make sure we don't accidentally expand big cubes!
+        if table[colname].ndim > 1:
+            if np.product(table[colname].shape[1:]) > 100:
+                raise ValueError(f"Table {colname} is too wide to expand into single columns: {table[colname].shape}")
         if table[colname].ndim == 2:
             for i in range(table[colname].shape[1]):
                 table[f'{colname}[{i}]'] = table[colname][:, i]
@@ -47,26 +54,33 @@ def table_to_glue_data(table, label_prefix, data_desc_id):
 
     kwargs = dict((c, table[c]) for c in table.colnames)
 
-    data = Data(label=label_prefix + f' [DATA_DESC_ID={data_desc_id}]', **kwargs)
+    data = Data(label=label, **kwargs)
 
     return data
 
 
-@data_factory(label='CASA Measurement Set', identifier=is_casa_measurement_set)
-def read_casa_measurement_set(filename, **kwargs):
+@data_factory(label='CASA Measurement Set', identifier=is_casa_table)
+def read_casa_table(filename, **kwargs):
 
     casa_table = CASATable.read(filename)
-
-    # Extract just the DATA_DESC_ID which can be done as a single table if we
-    # get just that column
-    data_desc_ids = np.sort(np.unique(casa_table.as_astropy_table(include_columns=['DATA_DESC_ID'])['DATA_DESC_ID']))
 
     datasets = []
     label_prefix = os.path.basename(filename)
 
-    for data_desc_id in data_desc_ids:
-        table = casa_table.as_astropy_table(data_desc_id=data_desc_id)
-        datasets.append(table_to_glue_data(table, label_prefix=label_prefix, data_desc_id=data_desc_id))
+    if casa_table.has_data_desc_id():
+
+        # Extract just the DATA_DESC_ID which can be done as a single table if we
+        # get just that column
+        data_desc_ids = np.sort(np.unique(casa_table.as_astropy_table(include_columns=['DATA_DESC_ID'])['DATA_DESC_ID']))
+
+        for data_desc_id in data_desc_ids:
+            table = casa_table.as_astropy_table(data_desc_id=data_desc_id)
+            datasets.append(table_to_glue_data(table, label=label_prefix + f' [DATA_DESC_ID={data_desc_id}]'))
+
+    else:
+
+        table = casa_table.as_astropy_table()
+        datasets.append(table_to_glue_data(table, label=label_prefix))
 
     return datasets