+++ CHANGES.rst
@@ -6,7 +6,7 @@
 
 - Started implementing a generic CASA table reader. [#12, #14, #15, #22, #32]
 
-- Add a glue data factory for measurement sets.
+- Add a glue data factory for CASA tables.
 
 0.1 (2020-11-08)
 ----------------